+++ server/data/cw3-docs/bvc.md
@@ -1,0 +1,39 @@
+# Baremetal Validators' Cone (BVC)
+
+## Wake up, Neo...
+
+Welcome, brave adventurer, to a decentralized era. You have a choice: continue renting centralized servers or take the red pill, seize control of your infrastructure, and operate on bare metal.
+
+## Help me, Obi-Wan Kenobi, you're my only hope.
+
+True decentralization means having physical access to your own hardware, owning your setup, and knowing it inside out. Bare metal node operation is essential for blockchain integrity and network security. Operators are guardians of distributed systems.
+
+## One ring to rule them all? Not today!
+
+Single points of failure in centralized cloud infrastructure are a risk. The answer is dispersing power across many hands, many locations, and many nodes. Bare metal validators are custodians of their own keys and infrastructure.
+
+## It's a dangerous business, Frodo, going out your door...
+
+Maintaining validator hardware presents challenges but the endeavor is worthwhile. This guide is a companion resource for navigating the landscape of bare metal validators, from the initial setup to maintaining a secure and reliable node.
+
+## Getting Started: Follow the White Rabbit
+
+Greetings, young Padawan. Much like Luke Skywalker taking his first steps into a larger world, starting your journey as a baremetal validator is an adventure filled with potential. Whether you're a seasoned pilot of the technical realm or just beginning to learn the ways of the Force, these guides will help you master the skills needed to bring true decentralization to the galaxy.
+
+In the pages ahead, we'll guide you through the initial stages of your mission. From selecting the right hardware to understanding the intricacies of network setup, you'll be equipped with all the tools you need to build and maintain your own node, becoming a true Jedi of the decentralized world.
+
+## Basics
+
+- Select your Router
+- Select your Switch
+- Select your Server
+- Networking
+- Monitoring
+- U.P.S. (Uninterruptible Power Supply)
+- P.D.U. (Power Distribution Unit)
+- Cables and co...
+- Server Cabinets
+
+## Join the Community
+
+Join the [Baremetal Validators' Cone](https://bvc.citizenweb3.com/) to step toward true decentralization without intermediaries.
