+++ server/data/cw3-docs/community.md
@@ -1,0 +1,29 @@
+# Web3 Society
+
+> [Join W3S](https://t.me/web_3_society)
+
+*Web3 Community Based on Values of Decentralization. Learning to Forget Tribalism and Love the Tech. Blockchain Agnostic. Fck Tribalism.*
+
+A digital space where you can share and find information about crypto technology, latest developments in consensus algorithms, DeAI news, open source hardware, software innovations, memes, guides, and make connections.
+
+To suggest adding a doc, open an [issue](https://github.com/citizenweb3/library/issues), submit a [PR](https://github.com/citizenweb3/library/pulls).
+
+## W3S docs
+
+1. [Web3 Squire documentation](https://github.com/citizenweb3/web3-society/tree/main/ambassadors)
+2. [Web3 libraries for various tools](https://github.com/citizenweb3/web3-society/blob/main/docs/web3%20libs%20for%20tools.md)
+3. [Web3 public API/RPC endpoints](https://github.com/citizenweb3/web3-society/blob/main/docs/web3%20public%20api.md)
+
+## The list
+
+- [Baremetal Validator Handbook](https://bvc.citizenweb3.com/)
+- [How to Build a Community: Guide for Startups](https://tribally.app/guide/)
+- [The Inside Story of the CryptoKitties Congestion Crisis](https://consensys.net/blog/news/the-inside-story-of-the-cryptokitties-congestion-crisis/)
+- [The Promise and Paradox of Decentralization](https://urbit.org/blog/2021-11-18-the-promise-and-paradox-of-decentralization)
+- [2021 Yearly Cryptocurrency Report](https://www.coingecko.com/buzz/2021-yearly-cryptocurrency-report)
+- [What Is APY, APR, and Impermanent Loss?](https://www.coingecko.com/buzz/what-is-apy-apr-and-impermanent-loss)
+- [DAO Governance Lessons Thus Far](https://www.coingecko.com/premium/publications/dao-governance-lessons-thus-far)
+- [Utopia of the anarchy of freedom](https://serejandmyself.github.io/blog/2019/08/07/The-anarchy-of-freedom.html)
+- [Cryptowiki](https://github.com/serejandmyself/cryptowiki/blob/master/cryptowiki.md)
+- [Learn How to Build a Crypto Company](https://a16z.com/crypto-startup-school/)
+- [A16Z blog](https://future.a16z.com/)
