+++ server/data/cw3-docs/manifesto.md
@@ -1,0 +1,17 @@
+# Citizen Web3 Manifesto
+
+"From this distant vantage point, the Earth might not seem of any particular interest. But for us, it's different. Consider again at that dot. That's here. That's home. That's us."
+
+The manifesto emphasizes humanity's place in the cosmos and warns against imagined self-importance and the delusion of privilege in the universe.
+
+## Mission
+
+Citizen Web3 is a non-custodial, self-hosted infrastructure provider and active member of the web3 galaxy. We are blockchain agnostic, censorship-resistant, and value-driven by the goals of cypherpunks.
+
+## Values and Commitments
+
+- Building off-the-grid infrastructure in the Atlantic Ocean
+- Prioritizing privacy networks
+- Believing in decentralization, cross-ecosystem collaboration, and technological sovereignty
+- Founded in 2020 (formerly Citizen Cosmos)
+- Mission to expand Web3 through infrastructure creation, knowledge sharing, and community empowerment -- without tribalism
