+++ server/data/cw3-docs/validator.md
@@ -1,0 +1,27 @@
+# Citizen Web3 Validator
+
+Located in the middle of the Atlantic ocean with off the grid capabilities.
+
+## Staking with Citizen Web3 Validator
+
+Citizen Web3 is a self-hosted, independent and non-custodial validator. Off the grid, baremetal infrastructure provider, with a focus on Privacy Networks.
+
+We work on a mixture of broadband and Starlink internet connections, and partially run off solar energy. We strongly believe in decentralized infrastructure, independent of the grid, and that's our end goal. Our full architecture is available in documentation.
+
+We use Multi-party computation software (Horcrux) on all our nodes, to protect stakers from double signing. We use Re-Stake on our mainnet nodes, set to restake twice per day, so users can compound their staking rewards more efficiently.
+
+## Features
+
+- Green Energy
+- Open Source on GitHub
+- Key Sharding via Horcrux
+- Auto Reward Re-Stake x2 per day
+
+## Services
+
+- [Staking](https://staking.citizenweb3.com/)
+- [ValidatorInfo](https://validatorinfo.com/)
+- [Podcast](https://www.citizenweb3.com/podcast)
+- [W.3.S. (Web3 Society)](https://t.me/web_3_society)
+- [B.V.C. (Baremetal Validators' Cone)](https://bvc.citizenweb3.com/)
+- [Chain Data Indexer](https://github.com/citizenweb3/chain-data-indexer)
