+++ src/app/services/AGENTS.md
@@ -28,6 +28,7 @@
 | `aztec-signal-service.ts` | Aztec signal events (getPayloadSignals, getPayloadSignalStats, isPayloadSubmitted, getPayloadSignalers, getSignalerStats, getOriginalPayload) |
 | `aztec-governance-service.ts` | Aztec governance display (getGovernanceConfigDisplay, getVotingPowerDisplay, getRawGovernanceConfig, getRawTotalVotingPower) |
 | `aztec-governance-db.ts` | Aztec governance data reader (getAztecGovernanceDataFromDb, StoredGovernanceConfig, AztecGovernanceData types) |
+| `podcast-service.ts` | Podcast RAG: vector search over podcast chunks (pgvector), episode summary lookups by validator identity/moniker, batch operations |
 | `redis-cache.ts` | Redis caching utilities (cacheGet, cacheSet, cacheGetOrFetch, CACHE_KEYS, CACHE_TTL) |
 
 ## Dependencies
@@ -79,7 +80,12 @@ src/app/services/
 ├── aztec-signal-service.ts   # Aztec signal events for governance
 ├── aztec-governance-service.ts # Aztec governance proposals
 ├── aztec-governance-db.ts    # Aztec governance DB helper
-└── redis-cache.ts            # Redis caching utilities
+├── podcast-service.ts        # Podcast RAG: vector search, episode summaries
+├── redis-cache.ts            # Redis caching utilities
+│
+├── ai/                       # AI chat assistant (see ai/AGENTS.md)
+│   ├── ai-service.ts         # LLM config, system prompt
+│   └── tools/                # LLM tools (chain, validator, governance, market, podcast)
 ```
 
 ## Common Patterns
